Client: inventory management: re-organize folders
