+++ packages/client/src/app/inventory-management/inventory-tools/article.ts
