+++ packages/client/src/app/inventory-management/inventory-tools/articles/articles.component.scss
