+++ packages/client/src/app/inventory-management/inventory-tools/home-page/home-page.component.scss
