+++ packages/client/src/app/inventory-management/inventory-tools/instant-stock/instant-stock.component.html
