+++ packages/client/src/app/inventory-management/inventory-tools/inventory-tools-routing.module.ts
@@ -16,4 +16,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class InvoiceToolRoutingModule {}
+export class InventoryToolsRoutingModule {}
