+++ packages/client/src/app/inventory-management/inventory-tools/invoices/invoices.component.scss
