+++ packages/client/src/app/inventory-management/inventory-tools/options-dialog/options-dialog.component.html
