+++ packages/client/src/app/inventory-management/inventory-tools/services/articles.service.ts
