+++ packages/client/src/app/inventory-management/inventory-tools/services/parse.service.ts
