+++ packages/client/src/app/inventory-management/providers/edit/edit.component.html
