+++ packages/client/src/app/inventory-management/providers/list/list.component.scss
