+++ packages/client/src/app/inventory-management/providers/new/new.component.html
