+++ packages/client/src/app/inventory-management/providers/new/new.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { DynamicFormModel, DynamicFormService } from '@ng-dynamic-forms/core';
 import { FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
-import { ProvidersService } from '../../../api-services/providers.service';
-import { ToasterService } from '../../../../core/services/toaster.service';
+import { ProvidersService } from '../../api-services/providers.service';
+import { ToasterService } from '../../../core/services/toaster.service';
 import { getProviderFromForm, getProviderModel } from '../providers.form-model';
 
 @Component({
