+++ packages/client/src/app/inventory-management/providers/provider.model.ts
