+++ packages/client/src/app/inventory-management/providers/providers-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { ProviderDetailResolverService } from '../../api-services/provider-detail-resolver.service';
+import { ProviderDetailResolverService } from '../api-services/provider-detail-resolver.service';
 import { NgxPermissionsGuard } from 'ngx-permissions';
 import { EditComponent } from './edit/edit.component';
 import { ViewComponent } from './view/view.component';
