+++ packages/client/src/app/inventory-management/providers/providers.data-source.ts
@@ -1,8 +1,8 @@
 import { Provider } from './provider.model';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { ProvidersService } from '../../api-services/providers.service';
+import { ProvidersService } from '../api-services/providers.service';
 import { CollectionViewer, DataSource } from '@angular/cdk/collections';
-import { MoleculerListOptions } from '../../../shared/models/MoleculerWrapper';
+import { MoleculerListOptions } from '../../shared/models/MoleculerWrapper';
 
 export class ProvidersDataSource implements DataSource<Provider> {
   private providersSubject = new BehaviorSubject<Provider[]>([]);
