+++ packages/client/src/app/inventory-management/providers/providers.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { NgxPermissionsModule } from 'ngx-permissions';
 
 import { ProvidersRoutingModule } from './providers-routing.module';
-import { SharedModule } from '../../../shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 import { NewComponent } from './new/new.component';
 import { EditComponent } from './edit/edit.component';
 import { ListComponent } from './list/list.component';
 import { ViewComponent } from './view/view.component';
-import { InventoryManagementApiServicesModule } from '../../api-services/inventory-management-api-services.module';
+import { InventoryManagementApiServicesModule } from '../api-services/inventory-management-api-services.module';
 
 @NgModule({
   declarations: [
