+++ packages/client/src/app/inventory-management/providers/view/view.component.html
